dev_requirements.txt
pytest==4.3.1
pytest-dotenv==0.4.0
vcrpy==2.0.1